requirements.txt
# Core Dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.10.0
soundfile>=0.12.0

# Audio Processing
pydub>=0.25.1
webrtcvad>=2.0.10
pyworld>=0.3.0
pyaudio>=0.2.11
noisereduce>=2.0.0
sounddevice>=0.4.6

# ML & Signal Processing
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# TTS Integration
piper-tts>=1.0.0
onnxruntime>=1.14.0
transformers>=4.38.0
accelerate>=0.27.0

# API & Web
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0

# CLI & Utilities
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
